# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: antsxdp2.0 admin  接口
  version: 2.0.0
  title: antsxdp 2.0 -admin
basePath: /antsxdp
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /sys/cdnsys/ip/control/list:
    post:
      tags:
        - IP管理
      summary: ip列表
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          type: string
          description: "control:-1=所有,0=未定义,1=白名单（7层）,2=7层黑名单,3=3层黑名单"
          default: "{\"parentId\":0,\"page\":1,\"limit\":10,\"control\":-1,\"ip\":\"\"}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/cdnsys/ip/control/save:
    post:
      tags:
        - IP管理
      summary: ip（分组） 保存与修改
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          type: string
          description:
          default: "{\"id\":0,\"parentId\":0,\"ip\":\"1.1.1.1\",\"control\":1,\"status\":0,\"remark\":\"\",,\"ips\":\"1.1.1.1,1.2.2.2\"}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/cdn/rewrite/list:
    post:
      tags:
        - URL转发
      summary: 列表
      produces:
        - application/json
      parameters:
        - name: body
          description: json
          in: body
          type: string
          default: "{\"user\":\"demo\",\"page\":1,\"limit\":20}"
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/cdn/rewrite/info/{id}:
    get:
      tags:
        - URL转发
      summary: 获取信息
      produces:
        - application/json
      parameters:
        - name: id
          description: 参数ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/cdn/rewrite/save:
    post:
      tags:
        - URL转发
      summary: 保存
      produces:
        - application/json
      parameters:
        - name: body
          description: json
          in: body
          type: string
          default: "{\"alias\":\"alias,alias\",\"id\":0,\"remark\":\"remark\",\"scheme\":\"http\",\"rewriteMode\":0,\"serialNumber\":\"1601213213\",\"serverName\":\"serverName\",\"status\":0,\"target\":\"string\",\"userId\":0}"
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/cdn/rewrite/delete:
    post:
      tags:
        - URL转发
      summary: 删除
      produces:
        - application/json
      parameters:
        - name: body
          description: json
          in: body
          type: string
          default: "{\"ids\":\"1,2,3\"}"
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'














        #定义数据模型
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...